clang warnings/errors

+++ trunk/linbox/tests/test-commentator.C
@@ -6,20 +6,20 @@
  *
  * --------------------------------------------------------
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -78,7 +78,7 @@ void runTestActivity (bool reportStrings)
 	if (reportStrings) outputReportStrings ();
 
 	for (int i = 0; i < 2; ++i) {
-		commentator().startIteration (i);
+		commentator().startIteration ((unsigned int)i);
 		if (reportStrings) outputReportStrings ();
 
 		commentator().start ("Special function 1", "function1");
