+++ trunk/linbox/tests/test-ntl-ZZ_p.C
@@ -75,7 +75,7 @@ int main (int argc, char **argv)
 	// Make sure some more detailed messages get printed
 	commentator().getMessageClass (INTERNAL_DESCRIPTION).setMaxDepth (2);
 
-	if (!runFieldTests (F, "NTL_ZZ_p", iterations, n, false)) pass = false;
+	if (!runFieldTests (F, "NTL_ZZ_p", (unsigned int)iterations, n, false)) pass = false;
 
 #if 0
 	FieldArchetype K(new NTL_ZZ_p(101));
